Fix publish target in Makefile
The existing promote target won't work because the datetime variable is
always going to be different from when the image was tagged using the
publish target. Also, pulling the stage image and then tagging and
pushing the prod image should all be run in the same target, otherwise
you cannot guarantee that what you have tagged locally as prod is the
same as what is currently in staging.

This commit combines what was the publish and promote targets into a
single publish target.