Update readme
Why these changes are being introduced:

Some of the links in the readme are not formatted properly, and we are
missing links to official and local docs.

Relevant ticket(s):

* https://mitlibraries.atlassian.net/browse/ETD-231

How this addresses that need:

This fixes the links and adds links to the docs.

Side effects of this change:

None.