Prep Matomo for AWS Org
Why these changes are being introduced:
We need to migrate our Matomo system from the legacy AWS to our new
AWS Organization. This requires a new design and an update to the
container. All the changes necessary to the container are in this
commit/branch.

How this addresses that need:
* Add a files folder to hold additional content that goes in the
custom container
* Include EnvironmentVariables v3.0.0 (for the Matomo 3 container) and
v4.0.1 (for the future Matomo 4 container)
* Update the config.ini.php source file
* Create a bash script to load into cron.daily for nightly backups of
core files
* Update Dockerfile to handle the plugin installation
* Update Dockerfile to handle the additional cron.daily job
* Update Dockerfile to use an explicit version of the Matomo Docker
container
* Update the README with detailed instructions for migration

Side effects of this change:
None. This branch should only be deployed in the new AWS Org.

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/IN-496
* https://mitlibraries.atlassian.net/browse/IN-487
* https://mitlibraries.atlassian.net/browse/IN-525