Upgrade to Matomo v4.x
Why these changes are being introduced:
Once deployed in the AWS Org, we will need to update from Matomo 3.x to
Matomo 4.x. This in turn means that we will need to update the plugins
in the container.

How this addresses that need:
* Remove the unused global.ini.php file
* Update Dockerfile with new version of EnvironmentVariables
* Update Dockerfile with new version of official Matomo Docker container
source
* Update config.ini.php to remove the incompatible pluging
(CustomPiwik.js) from the configuration (since it doesn't exist in the
Docker container)
* Add CustomVariables plugin to the `files/` folder
* Add Provider plugin to the `files/` folder
* Update Dockerfile to copy the updated CustomVariables and Provider
plugins to the container image
* Remove the Provider plugin from the list of installed plugins
* Clean up the backup-data.sh script
* Update the config.ini.php file
* Include the necessary .htaccess files (that should be automatically
generated by the Matomo install, but aren't) to address the "some
directories are publicly accessible warning"
* Update README.md
* Update the restore-from-backup HowTo

Side effects of this change:
None.

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/IN-533
* https://mitlibraries.atlassian.net/browse/IN-534