Switch to shell script for persistent data backup
Why these changes are being introduced:
The earlier attempt to schedule copies of some key files to the EFS
file system didn't work. This shifts the commands into a shell script
and this will require an update to the infrastructure that schedules
the copy command.

How this addresses that need:
* Create a shell script for the list of /bin/cp commands
* Update the Dockerfile to add the shell script to the /usr/local/bin
directory in the container

Side effects of this change:
None.