Fixes for renamed repository
Why these changes are being introduced:
With the renaming of this repository from "matomo" back to
"docker-matomo" there were a number of small modifications that needed
to be completed. This also includes a fix for an error running the
`docker build` command in a GitHub runner as well as a documentation
cleanup.

How this addresses that need:
* Update Makefile with corrected output from mitlib-tf-workloads-ecr
* Update dev-build workflow to match correct ECR repository and correct
OIDC-related IAM Role
* Update stage-build workflow to match correct ECR repository and
correct OIDC-related IAM Role
* Update prod-promote workflow to match correct ECR repository and
correct OIDC-related IAM Role
* Separate the single HOWTO.md into task-specific how-to documents
* Updates to README
* Move the HOWTOs into a HowTos subfolder in the docs folder

Side effects of this change:
None

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/IN-487

Side effects of this change:
None.