Upgrade to Matomo 4.x
Why these changes are being introduced:
The current version of Matomo (3.16.1) is woefully out-of-date. This
upgrades to the latest container version (4.12.0) and contains the
additional updates for the included plugins to match the new version of
Matomo.

How this addresses that need:
* Update Dockerfile with the latest version of Matomo from Docker Hub.
* Update the Matomo Upgrade HowTo with a few more details.

Side effects of this change:
After updating the container image in the ECR repository, it will need
to be deployed to the ECS Service via the AWS Console. Once the new
container is deployed, the database upgrade command will need to be
run via an "ssh" session to the new container. Once the database update
is completed, one more command will need to be run from the "ssh"
session on the container to update the tables to the `utf8mb4`
character set.

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/IN-537