matomo upgrade to 4.15.1
Why these changes are being introduced:

A new version of Matomo was released (v4.15.1). This change updates the
version of the official Matomo docker container that we derive our
container from resulting in an upgrade of our version of Matomo to
v4.15.1 from 4.13.1.

How this addresses that need:

A version increment in the Dockerfile

Side effects of this change:

Matomo is upgraded. Note: A database migration is not required as
part of this upgrade.

Changes to be committed:
      modified:   Dockerfile