Update Plugins
Why these changes are being introduced:
The recent upgrade to Matomo 5.0.1 failed to also address the
necessary updates to the two plugins (EnvironmentVariables and
CustomVariables) for this version of Matomo. This catches up the plugin
versions to match the Matomo version. During the recent database
resize, it was discovered that there were warnings in the Matomo web UI
about out-of-date plugin versions.

How this addresses that need:
* Add the new EnvironmentVariables-5.0.0 plugin to the repository
* Add the new CustomVariables-5.0.2 plugin to the repository
* Update the Dockerfile to use the two new plugin versions

Side effects of this change:
None.