Upgrade to Matomo 5.2.2
Why these changes are being introduced:
UXWS requested an upgrade to the latest version of Matomo. As part of
the initial testing of the upgrade in Dev1, it was noted that an
additional PlugIn appeared in the config.ini.php file.

How this addresses that need:
* Upgrade the Docker source from 5.1.2 to 5.2.2
* Update the config.ini.php file to include the newly listed
FeatureFlags PlugIn
* Update the config.ini.php to include the collation detail

Side effects of this change:
None.

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/INFRA-506
* https://mitlibraries.atlassian.net/browse/INFRA-499
* https://mitlibraries.atlassian.net/browse/INFRA-495