Upgrade Two PlugIns
Why these changes are being introduced:
During the testing and verification of the new 5.2.2 version of Matomo,
it was discovered that there were newer versions of the two extra
PlugIns that we use. At the same time, we cleared out some old copies
of plugins from this repository.

How this addresses that need:
* Add new CustomVariables 5.0.4 plugin folder
* Add new EnvironmentVariables 5.0.3 plugin folder
* Remove extra older plugin folders for both CustomVariables and
EnvironmentVariables plugins

Side effects of this change:
None.

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/INFRA-506