Install New Premium Plugins
Why these changes are being introduced:
The UXWS team purchased two premium plugins, the HeatmapSessionRecording
plugin and the UsersFlow plugin. The license key has already been
installed in the database, so we can proceed with the installation of
the plugins (but not the activation yet).

How this addresses that need:
* Unzip the two plugin packages into the `files/` folder in this
repository
* Update the `Dockerfile` to copy the source packages into the plugins
directory in the container
* Update the `config.ini.php` file to include the two new plugins
* Delete the leftover old versions of the two other non-premium plugins
that are part of our Matomo deployment

Side effects of this change:
None.

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/INFRA-517