Fix GHA Permissions
Why these changes are being introduced:
The workflows do not need to permission to make any changes to the code
itself, so we restrict to just `read-only`.

How this addresses that need:
* Add a line to each of the GHA workflows to restrict the workflow to
just read-only permissions to the code in the repository.
* Update the permissions for the job to allow GitHub Actions to
interact with the OIDC endpoint

Additionally,
* Add a `temp/` directory to the .gitignore file so that it's possible
to do some local work moving files around without worrying about
content getting accidentally pushed to GitHub
* Minor formatting changes to a few files

Side effects of this change:
None.