Documentation Updates
Why these changes are being introduced:
After much testing, this captures the latest understanding of how to
manage the installation of new premium plugins (e.g., those plugins
that require a license key and that might make changes to the database).

How this addresses that need:
* Minor edits to old migration document
* Update the How-To document for installing new premium plugins based
on the experience in Dev1 and Stage-Workloads for the current attempt

Side effects of this change:
None.

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/INFRA-517