Phase 2 of Premium Plugins Installation
Why these changes are being introduced:
Phase 1 of the premium plugins installation is complete across
Dev1, Stage-Workloads, and Prod-Workloads (that is, the plugin folders
are installed in the container and the plugins are listed in the
Plugin Directory in the UI, but they are not activated). The plugins
have been manually activated in Stage & Prod, so any database schema
changes are complete. As part of the testing, InnoCraft support let us
know that the `CustomJsTracker` plugin (a Core plugin) must be activated
for the HeatmapSessionRecording and the UsersFlow plugins to work
correctly (because both plugins actually make changes to the matomo.js
and piwik.js files).

How this addresses that need:
* Update the Dockerfile to include the updated `matomo.js and `piwik.js`
files (that are the result of activating the plugins); these files are
written to the `/usr/src/matomo/` directory because the `entrypoint.sh`
script is what actually creates the `/var/www/html/` directory when the
container is launched
* Update the `config.ini.php` file to correctly list all the plugins
that exist in the `plugins/` directory and sort them alphabetically to
make it easier to review them in the future
* Update the `config.ini.php` file to include the `CustomJsTracker` as
an activated plugin
* Update the `config.ini.php` file to include the UsersFlow and
HeatmapSessionRecording plugins as active plugins
* Include the version of the `matomo.js` and `piwik.js` files that are
generated by the new plugins

Side effects of this change:
None.
 Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/INFRA-517