Update GHA GITHUB_TOKEN Permissions
Why these changes are being introduced:
There is a potential conflict between GITHUB_TOKEN permissions declared
in a caller workflow (like this) and the shared workflow that is being
called. After testing various combinations, the end result is as simple
as setting the correct GITHUB_TOKEN permissions in the shared workflow
and not declaring any permissions in the calling workflow. If there is
some need to declare GITHUB_TOKEN permissions in the calling workflow,
they MUST match the permissions in the shared workflow.

Since checkov will throw a warning if no permissions are declared at all
there is a checkov:skip comment in the caller workflow. Sadly, for an
unknown reason, checkov doesn't register this skip comment and still
shows a warning.

How this addresses that need:
* Remove all permissions declared in the caller workflows
* Add a checkov:skip comment in the workflows, even though it doesn't
properly work (this at least lets any future developer know that there
is a good reason for not having a permissions statement in the yaml)

Side effects of this change:
None.

Related Jira Tickets:
* https://mitlibraries.atlassian.net/browse/IR-238