TEMPORARY: Update Workflows
Why these changes are being introduced:
We are using this docker-matomo repository to verify that the updates
to the shared workflows for our automated deployment are working
properly. Once this is all the way through to Prod, we'll merge the
PR on the .github repo and then go through this again to clean up the
workflows.

How this addresses that need:
* Update the workflows to point to a branch in the .github repository
to use updated versions of our shared deployment workflows