Update to Matomo 5.3.3 and Fix TagManager
Why these changes are being introduced:
Along with the 5.3.3 upgrade, we are enabling the TagManager plugin.
We just discovered that the TagManager plugin had been enabled via
the Matomo UI in Prod, but since this was never captured in the
container definition here, it did not persist.

NOTE: The steps documented in the HOWTO have already been followed in
Dev1, Stage-Workloads, and Prod-Workloads.

How this addresses that need:
* Create a HOWTO to document the difficulty in enabling the TagManager
plugin in our environment
* Update the config.ini.php file to reflect the activation of the
TagManager plugin
* Update the Dockerfile to update to the latest version of Matomo

Side effects of this change:
None.

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/WEB-2064
* https://mitlibraries.atlassian.net/browse/INFRA-526