Updates to match new infrastructure
Why these changes are being made:
With the updates to the infrastructure and the use of peristent storage
for a few of the paths in the `/var/www/html`, we needed to change
how we managed the container.

How these changes are implemented:
* Update CODEOWNERS to force review by Dhruti when the Dockerfile is
modified
* Minor adjustments to the config.ini.php file related to report
archiving.
* Update the backup script to use tar (this is just copying the
techinque I found in the entrypoint script)
* Minor updates to some documentation

Related Jira Tickets:
* https://mitlibraries.atlassian.net/browse/IN-1301