Add New Premium Plugin
Why these changes are being introduced:
A third premium plugin was purchased from InnoCraft. At the same time,
one of the other premium plugins had a pending update. Per the
documentation in the HOWTO-premium-plugins.md, this pass through the
repository just installs the plugin directory and files but does not
activate the plugin.

Additionally, we revert the copying of the config.ini.php file back to
the /var/www/html folder instead into the source folder. Now that we
have established the persistent storage for the config.ini file, we
don't need to worry about forcing it into the source.

How this addresses that need:
* Add the source code for the SearchEngineKeywordsPerformance plugin to
the files/ directory
* Add the new version for the HeatmapSessionRecording plugin to the
files/ directory
* Update the Dockerfile to copy the new plugin sources to the plugins/
directory
* Update the Dockerfile to copy the config.ini.php to /var/www/html
instead of to /usr/src/matomo
* Update the backup-data.sh script to make it a little cleaner
* Update the HOWTO-premium-plugins.md documentation
* Update the HOWTO-miscellaneous.md documentation to address a TooManyRedirectsException

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/INFRA-558