Activate SearchEngineKeywordsPerformance Plugin
Why these changes are being introduced:
The files are in place in the container across all three environments
and the `./console plugin:activate` command has been run once in all
three environments (which preps the database). Now we need to ensure
that any container redeploys maintain the active status for this plugin.

How this addresses that need:
* Add the SearchEngineKeywordsPerformance to the list of installed
plugins
* Add the SearchEngineKeywordsPerformance to the list of active plugins

Side effects of this change:
None.

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/INFRA-558