Upgrade to Matomo 5.4.0
Why these changes are being introduced:
Upgrade the Docker container to run Matomo 5.4.0 so that we can enable
the new CustomReports plugin.

How this addresses that need:
* Update the Dockerfile to pull the 5.4.0 container UrlSource
* Update the Dockerfile to add the CustomReports plugin contents to the
container (but do not install or activate it)
* Update the HowTo information for premium plugins to make it a little
more clear how to get started with a new premium plugin
* Add the other plugin updates to the repository but do not add them
to the container yet

Side effects of this change:
None.

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/INFRA-577
* https://mitlibraries.atlassian.net/browse/INFRA-576
* https://mitlibraries.atlassian.net/browse/INFRA-569