Installation of Custom Reports Plugin
Why these changes are being introduced:
UXWS Requested the installation of the Custom Reports premium plugin.

The prep work has already been completed (loading the files into the
container). Now it's time to move on to enabling the plugin and
ensuring that the container definition is up-to-date.

How this addresses that need:
* Update the config.ini.php file to match the changes that appeared
in the running container

Side effects of this change:
None.

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/INFRA-576